+++ js/canvasRenderer.js
@@ -1,4 +1,4 @@
-"use strict";
+﻿"use strict";
 
 function clamp(number, minimum, maximum) {
     return Math.min(Math.max(number, minimum), maximum);